Organize main site images
