+++ src/components/MoreClients.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Web from '../../static/img/web.svg';
-import Desktop from '../../static/img/monitor.svg';
+import Web from '../../static/images/icons/web.svg';
+import Desktop from '../../static/images/icons/monitor.svg';
 import {
   Android,
   Apple,
@@ -8,7 +8,7 @@ import {
   Amazon,
   Kodi
 } from '@icons-pack/react-simple-icons';
-import Plus from '../../static/img/plus-thick.svg';
+import Plus from '../../static/images/icons/plus-thick.svg';
 
 export default function MoreClients() {
   return (
